Fix warnings

+++ EnvironmentSimulator/Modules/ScenarioEngine/SourceFiles/Entities.cpp
@@ -1026,8 +1026,8 @@ Object::OverlapType Object::OverlappingFront(Object* target, double tolerance)
 	// At least one vertex on each side of front line: OVERLAP_FULL
 
 	// Own object front side of bounding box
-	SE_Vector front_left(boundingbox_.center_.x_ + boundingbox_.dimensions_.length_ / 2.0, boundingbox_.dimensions_.width_ / 2.0);
-	SE_Vector front_right(boundingbox_.center_.x_ + boundingbox_.dimensions_.length_ / 2.0, -boundingbox_.dimensions_.width_ / 2.0);
+	SE_Vector front_left(boundingbox_.center_.x_ + boundingbox_.dimensions_.length_ / 2.0f, boundingbox_.dimensions_.width_ / 2.0f);
+	SE_Vector front_right(boundingbox_.center_.x_ + boundingbox_.dimensions_.length_ / 2.0f, -boundingbox_.dimensions_.width_ / 2.0f);
 
 	// Rotate and translate front line
 	front_left = front_left.Rotate(pos_.GetH());
@@ -1038,10 +1038,10 @@ Object::OverlapType Object::OverlappingFront(Object* target, double tolerance)
 	// Specify target object bounding box corner vertices, starting at first quadrant going clock wise
 	SE_Vector vertex[4] =
 	{
-		{ target->boundingbox_.center_.x_ + target->boundingbox_.dimensions_.length_ / 2.0, target->boundingbox_.center_.y_ + target->boundingbox_.dimensions_.width_ / 2.0 },
-		{ target->boundingbox_.center_.x_ - target->boundingbox_.dimensions_.length_ / 2.0, target->boundingbox_.center_.y_ + target->boundingbox_.dimensions_.width_ / 2.0 },
-		{ target->boundingbox_.center_.x_ - target->boundingbox_.dimensions_.length_ / 2.0, target->boundingbox_.center_.y_ - target->boundingbox_.dimensions_.width_ / 2.0 },
-		{ target->boundingbox_.center_.x_ + target->boundingbox_.dimensions_.length_ / 2.0, target->boundingbox_.center_.y_ - target->boundingbox_.dimensions_.width_ / 2.0 }
+		{ target->boundingbox_.center_.x_ + target->boundingbox_.dimensions_.length_ / 2.0f, target->boundingbox_.center_.y_ + target->boundingbox_.dimensions_.width_ / 2.0f },
+		{ target->boundingbox_.center_.x_ - target->boundingbox_.dimensions_.length_ / 2.0f, target->boundingbox_.center_.y_ + target->boundingbox_.dimensions_.width_ / 2.0f },
+		{ target->boundingbox_.center_.x_ - target->boundingbox_.dimensions_.length_ / 2.0f, target->boundingbox_.center_.y_ - target->boundingbox_.dimensions_.width_ / 2.0f },
+		{ target->boundingbox_.center_.x_ + target->boundingbox_.dimensions_.length_ / 2.0f, target->boundingbox_.center_.y_ - target->boundingbox_.dimensions_.width_ / 2.0f }
 	};
 
 	for (int i = 0; i < 4; i++)  // for all vertices
@@ -1074,11 +1074,11 @@ Object::OverlapType Object::OverlappingFront(Object* target, double tolerance)
 		{
 			inside_count++;
 
-			if (s_norm * boundingbox_.dimensions_.width_ < tolerance)  // s_norm is factor (0..1) along front line
+			if (s_norm * static_cast<double>(boundingbox_.dimensions_.width_) < tolerance)  // s_norm is factor (0..1) along front line
 			{
 				outside_left_count++;
 			}
-			else if ((1 - s_norm) * boundingbox_.dimensions_.width_ < tolerance)
+			else if ((1 - s_norm) * static_cast<double>(boundingbox_.dimensions_.width_) < tolerance)
 			{
 				outside_right_count++;
 			}
@@ -1338,7 +1338,7 @@ Vehicle::Vehicle() : Object(Object::Type::VEHICLE), trailer_coupler_(nullptr), t
 
 Vehicle::Vehicle(const Vehicle& v) : Object(Object::Type::VEHICLE), trailer_coupler_(nullptr), trailer_hitch_(nullptr) // TODO: @Emil
 {
-	*this = v; // TODO: @Emil
+	*this = v;
 
 	if (v.trailer_coupler_ && v.trailer_coupler_->tow_vehicle_)
 	{
